+++ Doublylinkedlist.h
@@ -6,14 +6,14 @@
 //
 typedef struct Node
 {
- int number;  // or double 2^16 -1 
+ unsigned long long int number;  // 2^128 -1  .... max 19 position 
  struct Node *prev;
  struct Node *next;
 } Node;
 
 typedef struct Dlist
 {
- int size; // size of number
+ long int size; // size of number
  char sign;//? //sign of number
  Node *head; // [pointer on head
  Node *tail; // pointr on tail
@@ -23,12 +23,12 @@ typedef struct Dlist
 void init_Dlist (Dlist *BigNum);
 void destroy_pointer_Dlist (Dlist *BigNum);
 void destroy_fixed_Dlist (Dlist *BigNum);
-void insert_to_end_Dlist (Dlist *BigNum, int *number);
-void insert_to_begin_Dlist (Dlist *BigNum ,int *number);
+void insert_to_end_Dlist (Dlist *BigNum, unsigned long long int *number);
+void insert_to_begin_Dlist (Dlist *BigNum ,unsigned long long int *number);
 void delete_head (Dlist *BigNum);
 void delete_tail (Dlist *BigNum);
-void delete_Node_pointer_Dlist (Dlist *BigNum, int position);
-Node *get_Node_DList (Dlist* BigNum, int position);
-void insert_Node_Dlist (Dlist *BigNum, int position);
-int *get_Node_number_Dlist(Dlist *BigNum,int position);
+void delete_Node_pointer_Dlist (Dlist *BigNum, long int position);
+Node *get_Node_DList (Dlist* BigNum, long int position);
+void insert_Node_Dlist (Dlist *BigNum,  long int position);
+unsigned long long int  *get_Node_number_Dlist(Dlist *BigNum, long int position);
